Fix codenarc NoDouble rules : Checks for use of the double or Double types, in fields, variables, method parameters and method return types.

+++ applications/order/groovyScripts/order/OrderView.groovy
@@ -406,7 +406,7 @@ if (orderHeader) {
         if (productId) {  // avoid order items without productIds, such as bulk order items
             resultOutput = runService('getProductManufacturingSummaryByFacility', [productId: productId])
             manufacturingInQuantitySummaryByFacility = resultOutput.summaryInByFacility
-            Double productionQuantity = 0
+            BigDecimal productionQuantity = 0
             manufacturingInQuantitySummaryByFacility.values().each { manQuantity ->
                 productionQuantity += manQuantity.estimatedQuantityTotal
             }
